Move `assignments.ts` hooks to `/services`
Hooks that *use* RTK queries do not belong in `../state/slices`; this should be for slice definitions only.

Note that I have only moved files around in this commit, the contents of `../services/assignments.ts` has changed accordingly and Git can't comprehend this well
