+++ src/app/state/index.ts
@@ -53,7 +53,6 @@ export * from "./actions/quizzes";
 export * from "./middleware/notificationManager";
 export * from "./middleware/userConsistencyChecker";
 export * from "./slices/api/gameboards";
-export * from "./slices/api/assignments";
 export * from "./slices/admin";
 export * from "./slices/user";
 export * from "./slices/topic";
